pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cacheiro"
version = "0.1.0"
description = "GA-based TSP/VRP optimizer with LLM reporting for hospital deliveries"
authors = [{name = "Equipe"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "folium>=0.15",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "openai>=1.12",
    "streamlit>=1.32",
    "google-generativeai>=0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"
